Remove force flag
I thought I could force if I set `triggers.when_value_changed` to a
random value whenever I wanted to force. But I'd need to keep the same
value *unless* the force flag was enabled.

The right move for forcing regeneration is to use `terraform taint` on
the resource. That should tell terraform to rotate the resource out,
even if the state hasn't changed.